// src/web/src/api/generated/models/ActivitySourceResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { JsonElement } from './JsonElement';
export type ActivitySourceResponse = {
    id: string;
    sourceClass: string;
    sourceRowNumber: number | string | null;
    linkingResult: string | null;
    observedAt: string | null;
    summary: JsonElement;
    sourceFileId: string | null;
    originalName: string | null;
    sha256: string | null;
    ingestionRunId: string | null;
};